Extract shared coreml op registrations, fix subprocess and constants
- Extract duplicated torch op registrations (new_ones, new_zeros, multiply)
  from 3 files into scripts/coreml_ops.py
- Replace os.system with subprocess.run for mlmodelc compilation
- Import BUCKETS from export_coreml in stage_harness instead of hardcoding
