+++ scripts/stage_harness.py
@@ -28,32 +28,9 @@ nn.utils.rnn.pack_padded_sequence = lambda x, lengths, **kw: x
 nn.utils.rnn.pad_packed_sequence = lambda x, **kw: (x, None)
 
 import coremltools as ct
-from coremltools.converters.mil.frontend.torch.torch_op_registry import (
-    _TORCH_OPS_REGISTRY, register_torch_op,
-)
-from coremltools.converters.mil import Builder as mb
+from coreml_ops import register_missing_torch_ops
 
-if "new_ones" not in _TORCH_OPS_REGISTRY:
-    @register_torch_op
-    def new_ones(context, node):
-        size = context[node.inputs[1]]
-        shape = mb.cast(x=size, dtype="int32", name=node.name + "_shape")
-        result = mb.fill(shape=shape, value=1.0, name=node.name)
-        context.add(result)
-
-if "new_zeros" not in _TORCH_OPS_REGISTRY:
-    @register_torch_op
-    def new_zeros(context, node):
-        size = context[node.inputs[1]]
-        shape = mb.cast(x=size, dtype="int32", name=node.name + "_shape")
-        result = mb.fill(shape=shape, value=0.0, name=node.name)
-        context.add(result)
-
-if "multiply" not in _TORCH_OPS_REGISTRY:
-    @register_torch_op
-    def multiply(context, node):
-        from coremltools.converters.mil.frontend.torch.ops import mul
-        mul(context, node)
+register_missing_torch_ops()
 
 sys.path.insert(0, os.path.dirname(__file__))
 from export_coreml import patch_sinegen_for_export, SAMPLES_PER_FRAME
@@ -747,8 +724,10 @@ def main():
     print("Loading model...")
     pipeline, model, set_phases_fn = load_model()
 
-    max_tokens = 124
-    max_audio = 175_800
+    from export_coreml import BUCKETS
+    bucket = BUCKETS["kokoro_21_5s"]
+    max_tokens = bucket["max_tokens"]
+    max_audio = bucket["max_audio"]
     max_frames = max_audio // SAMPLES_PER_FRAME
 
     voice_pack = pipeline.load_voice("af_heart")
